% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreProteinFragments.R
\name{.chColNa}
\alias{.chColNa}
\title{Check Column Names from Matrix Or data.frame}
\usage{
.chColNa(x, colNa = "index", rmCol = NULL, callFrom = NULL)
}
\arguments{
\item{x}{(matrix or data.frame)}

\item{colNa}{(character)}

\item{rmCol}{(character)}

\item{callFrom}{(character) allow easier tracking of message(s) produced}
}
\value{
This function returns a matrix or data.frame with adjusted columns
}
\description{
Check matrix or data.frame for containing columns with specified name and optionally remove other columns (by their names)
}
\examples{
ma1 <- matrix(1:6, nrow=2, dimnames=list(NULL, c("index","aa","bb")))
.chColNa(ma1)
.chColNa(ma1, colNa="zz", rmCol="aa") 
}
\seealso{
\code{\link{scoreFragments}}
}
