% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPotModifAAs.R
\name{.countLET}
\alias{.countLET}
\title{Count Letters}
\usage{
.countLET(sequ, countCh = "K", silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{sequ}{(character) eg peptide sequence(s)}

\item{countCh}{(charcter)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) for bug-tracking: more/enhanced messages and intermediate objects written in global name-space}

\item{callFrom}{(character) allow easier tracking of message(s) produced}
}
\value{
This function returns a numeric vector of counts for 'countCh' (single element !) in each element of 'seq'
}
\description{
Count how many time a given letter occurs in each element of 'seq'
}
\examples{
protP2 <- c(mesp="MESPEPTIDES", pepe="PEPEPEP")
.countLET(protP2,"P")
}
\seealso{
\code{\link{AAfragSettings}}, \code{\link{makeFragments}}
}
