% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMassModif.R
\name{.multMatByColNa}
\alias{.multMatByColNa}
\title{Multiply Values Of Matrix By Its Colnames And Sum By Row}
\usage{
.multMatByColNa(
  mat,
  sumByRow = TRUE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{mat}{(matrix) main input}

\item{sumByRow}{(logical)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This functions returns a numeric vector or a matrix if \code{sumByRow=FALSE}
}
\description{
This function allows multiplying values of 'mat' by its colnames and (optionally) summing along rows.
}
\examples{
mat1 <- 3 + matrix(1:4, ncol=2, dimnames=list(letters[1:2], c("3","2")))
.multMatByColNa(mat1)
.multMatByColNa(mat1, sumByRow=FALSE)
}
\seealso{
\code{\link[wrMisc]{convToNum}}
}
