% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalIsoFragm.R
\name{.prefFragPattern}
\alias{.prefFragPattern}
\title{Return data.frame with pattern of perferential fragmentation sites}
\usage{
.prefFragPattern(silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a data.frame with pattern of perferential fragmentation sites
}
\description{
Return data.frame with pattern of perferential fragmentation sites
Here a simplified version (elaborate see Kelleher group: Haverland 2017, J Am Soc Mass Spectrom)
}
\examples{
peTab <- matrix(c("9","13","14","15", "LPVIAGHEAAG","PVIAGHEAAGI","EKKPFSI","KKPFSIE", 
  "P","L","E","E", "I","V","E","E"),nr=4,dimnames=list(NULL,c("no","seq","precAA","tailAA")))
head(.prefFragPattern())
}
\seealso{
\code{\link{makeFragments}}
}
