% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randMassByStochastic.R
\name{randMassByStochastic}
\alias{randMassByStochastic}
\title{Make Decoy Mass By Full Randomization}
\usage{
randMassByStochastic(
  xChar,
  nRepeat = 5,
  negAvoid = TRUE,
  setSeed = NULL,
  silent = FALSE,
  debug = TRUE,
  callFrom = NULL
)
}
\arguments{
\item{xChar}{(numeric vector) characterize main data, must conatain elements 'n','minV','maxV'}

\item{nRepeat}{(integer) number of time whole randomization process should be repeated}

\item{negAvoid}{(logical) if TRUE try avoiding 0 or negative random mass in result}

\item{setSeed}{(integer) seed for random number generation set.seed()}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages and objects exportet to current session for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a matrix with additional column 'decoyMass', or if 'sepCol'=FALSE as additional lines
}
\description{
Make full random decoy mass vector (mimick pepTab)
}
\examples{
rand <- randMassByStochastic(c(n=10, minV=2, maxV=7))
summary(rand)
}
\seealso{
\code{\link[stats]{Uniform}}, \code{\link{randMassByMut}}, \code{\link[wrMisc]{convToNum}}
}
